// src/modules/products/interfaces/data-product-erp-for-shipping.ts
export type dataProductErpForShipping = {
        codigo: number
        preco: number
        grupo: number
        controle_lote_serie: 'S' | 'N'
        unidade_medida: string
        descricao: string
        num_fabricante: string
        num_original: string
        sku: string
        outro_cod2:string
        marca: number
        ativo: 'S' | 'N'
        tipo: number
        class_fiscal: string
        origem: string
        cst: string
        data_cadastro: string
        observacoes1: string
        observacoes2: string
        observacoes3: string
        estoque:number
}
 